requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1;python_version<"3.9"
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dialogflow==2.0.0
feedparser==6.0.10
Flask==2.1.3
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.32.0
google-auth==1.35.0
google-cloud-dialogflow==2.15.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
